refactor(desktop): split app shell turn view model
