+++ apps/desktop/src/renderer/app-shell.tsx
@@ -23,15 +23,12 @@ import {
   ChatView,
   Composer,
   type ComposerHandle,
-  deriveTurnLineageMap,
   type MakaUriDest,
   MakaUriContext,
-  materializeTurns,
   type NavSelection,
   SessionListPanel,
   type SkillEntry,
   type TurnFooterActionMeta,
-  type TurnLineageBadge,
   useToast,
   type ToolActivityItem,
   activePermissionFor,
@@ -50,12 +47,10 @@ import { deriveChatHeaderAlert } from './chat-header-alert';
 import { deriveStaleSessionIds } from './stale-sessions';
 import { deriveSessionStatusGroups } from './session-status-grouping';
 import {
-  deriveFailedTurnRecovery,
-  describeTurnErrorClass,
   presentSessionStatus,
   sessionStatusAriaLabel,
 } from './session-status-presentation';
-import { deriveTurnFooterActions } from './turn-footer-actions';
+import { deriveAppShellTurnViewModel } from './app-shell-turn-view-model';
 import { readScrollMotionBehavior } from './scroll-motion-policy';
 import { deriveBranchBanner } from './branch-banner';
 import { pickCatalogDefaultChatModel } from './model-catalog-choices';
@@ -554,99 +549,21 @@ export function AppShell() {
     toastApi,
   });
 
-  // PR109e: per-turn auxiliary view-model. Combines:
-  //  - footer actions (PR109d) — status + lineage + pending
-  //  - failed reason label (PR109e-d) — errorClass → Chinese via
-  //    describeTurnErrorClass, NEVER exposes raw enum
-  //  - lineage badges (PR109e-e) — forward "重试自 turn X" on the new
-  //    turn + reverse "已重试 → turn Y" on the origin, derived from
-  //    deriveTurnLineageMap (which already exists in @maka/ui).
   const {
     turnFooterActionsByTurn,
     turnFailedReasonLabels,
     turnFailedRecoveryLabels,
     turnLineageBadgesByTurn,
-  } = useMemo(() => {
-    const turnsForLineage = materializeTurns(messages, liveTools);
-    const lineage = deriveTurnLineageMap(turnsForLineage);
-    const turnsById = new Map(turnsForLineage.map((t) => [t.turnId, t]));
-    const shortId = (turnId: string) => turnId.slice(0, 6);
-    const footer: Record<string, ReadonlyArray<TurnFooterActionMeta>> = {};
-    const failedLabels: Record<string, string> = {};
-    const failedRecoveryLabels: Record<string, string> = {};
-    const badges: Record<string, TurnLineageBadge[]> = {};
-    for (const turn of turnsForLineage) {
-      const lineageEntry = lineage.get(turn.turnId);
-      const pendingForTurn = new Set<TurnFooterActionMeta['id']>();
-      for (const id of ['retry', 'regenerate', 'branch', 'copy'] as const) {
-        if (activeId && pendingTurnActions.has(pendingKeyOf(activeId, turn.turnId, id))) {
-          pendingForTurn.add(id);
-        }
-      }
-      footer[turn.turnId] = deriveTurnFooterActions({
-        status: turn.status,
-        hasContent: Boolean(turn.assistant?.text && turn.assistant.text.trim().length > 0),
-        ...(lineageEntry?.retriedToTurnId ? { alreadyRetried: true } : {}),
-        ...(lineageEntry?.regeneratedToTurnId ? { alreadyRegenerated: true } : {}),
-        ...(pendingForTurn.size > 0 ? { pendingActions: pendingForTurn } : {}),
-      });
-      if (turn.status === 'failed') {
-        failedLabels[turn.turnId] = describeTurnErrorClass(turn.errorClass);
-        failedRecoveryLabels[turn.turnId] = deriveFailedTurnRecovery({
-          errorClass: turn.errorClass,
-          partialOutputRetained: turn.partialOutputRetained,
-          toolActivityCount: turn.tools.length,
-          erroredToolCount: turn.tools.filter((tool) => tool.status === 'errored').length,
-        }).label;
-      }
-      const turnBadges: TurnLineageBadge[] = [];
-      // Forward badges — pointing back at the origin
-      if (turn.retriedFromTurnId && turnsById.has(turn.retriedFromTurnId)) {
-        turnBadges.push({
-          id: `forward-retry-${turn.turnId}`,
-          label: `重试自 turn ${shortId(turn.retriedFromTurnId)}`,
-          tooltip: `这是对上一轮回答的重试`,
-          targetTurnId: turn.retriedFromTurnId,
-          direction: 'forward',
-        });
-      }
-      if (turn.regeneratedFromTurnId && turnsById.has(turn.regeneratedFromTurnId)) {
-        turnBadges.push({
-          id: `forward-regen-${turn.turnId}`,
-          label: `重新生成自 turn ${shortId(turn.regeneratedFromTurnId)}`,
-          tooltip: `保留旧回答，重新生成的并行回答`,
-          targetTurnId: turn.regeneratedFromTurnId,
-          direction: 'forward',
-        });
-      }
-      // Reverse badges — pointing at descendants (derived map)
-      if (lineageEntry?.retriedToTurnId && turnsById.has(lineageEntry.retriedToTurnId)) {
-        turnBadges.push({
-          id: `reverse-retry-${turn.turnId}`,
-          label: `已重试 → turn ${shortId(lineageEntry.retriedToTurnId)}`,
-          tooltip: `跳转到对此回答的重试`,
-          targetTurnId: lineageEntry.retriedToTurnId,
-          direction: 'reverse',
-        });
-      }
-      if (lineageEntry?.regeneratedToTurnId && turnsById.has(lineageEntry.regeneratedToTurnId)) {
-        turnBadges.push({
-          id: `reverse-regen-${turn.turnId}`,
-          label: `已重新生成 → turn ${shortId(lineageEntry.regeneratedToTurnId)}`,
-          tooltip: `跳转到对此回答的重新生成`,
-          targetTurnId: lineageEntry.regeneratedToTurnId,
-          direction: 'reverse',
-        });
-      }
-      if (turnBadges.length > 0) badges[turn.turnId] = turnBadges;
-    }
-    return {
-      turnFooterActionsByTurn: footer,
-      turnFailedReasonLabels: failedLabels,
-      turnFailedRecoveryLabels: failedRecoveryLabels,
-      turnLineageBadgesByTurn: badges,
-    };
-  }, [activeId, messages, liveTools, pendingTurnActions]);
+  } = useMemo(
+    () => deriveAppShellTurnViewModel({
+      activeId,
+      messages,
+      liveTools,
+      pendingTurnActions,
+      pendingKeyOf,
+    }),
+    [activeId, messages, liveTools, pendingTurnActions],
+  );
 
   // PR109e-e: click handler for lineage badge → scroll target turn into
   // view. Avoids pulling a separate ref-tracker: relies on the
